Description optimisation up to 1275.
"optimisation" here means shortening longish descriptions to be the minimum
necessary to sufficiently deliniate the concept.  This drops explanations
and remarks, in particular with a few to not confusing LLMs that might
use these descriptions to learn the concepts.

Where I changed the descriptions significantly enough to make the attribution
to the Heydary-Malayeri dictionary questionable, I removed the corresponding
editorial notes.  I'd suggest for replace them with the git history,
but I could imagine some hack to introduce editorial notes from the git history
(though I'm not wild on that at all).